refactor: set proving_system_id to i64 in task table struct

+++ aggregation_mode/batcher/src/db.rs
@@ -48,7 +48,7 @@ impl Db {
     pub async fn insert_task(
         &self,
         address: &str,
-        proving_system_id: i32,
+        proving_system_id: i64,
         proof: &[u8],
         program_commitment: &[u8],
         merkle_path: Option<&[u8]>,
